requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
langchain==0.0.340
langchain-openai==0.0.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
spacy==3.7.2
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
gunicorn==21.2.0
python-multipart==0.0.6
pydantic==2.4.2
sqlalchemy==2.0.21
sqlite3
jinja2==3.1.2
markdown==3.5.1
pyttsx3==2.90
speechrecognition==3.10.0
pyaudio==0.2.11
